Moved socket-io into an own library, and only include the compiled binaries in PomocCore

+++ Framework/Pomoc/PomocDriverApp/PomocDriverApp/ViewController.m
@@ -140,7 +140,7 @@
 }
 
 // Probably don't need to edit anything below
-- (int)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
+- (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
 {
     return self.messages.count;
 }
